Ajuste completo de las tablas con el nuevo modelo de la base de datos

+++ src/main/java/com/example/demo/model/Estado.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class Estado {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long estado_id;
+    private Integer estado_id;
 
     @Column(nullable = false, length = 100)
     private String estado;
